+++ oos-20190601/src/Models/ListApplicationGroupsResponseBody/applicationGroups.php
@@ -41,6 +41,16 @@ class applicationGroups extends Model
     /**
      * @var string
      */
+    public $errorDetail;
+
+    /**
+     * @var string
+     */
+    public $errorType;
+
+    /**
+     * @var string
+     */
     public $importTagKey;
 
     /**
@@ -74,6 +84,8 @@ class applicationGroups extends Model
         'deployParameters' => 'DeployParameters',
         'deployRegionId' => 'DeployRegionId',
         'description' => 'Description',
+        'errorDetail' => 'ErrorDetail',
+        'errorType' => 'ErrorType',
         'importTagKey' => 'ImportTagKey',
         'importTagValue' => 'ImportTagValue',
         'name' => 'Name',
@@ -112,6 +124,14 @@ class applicationGroups extends Model
 
         if (null !== $this->description) {
             $res['Description'] = $this->description;
+        }
+
+        if (null !== $this->errorDetail) {
+            $res['ErrorDetail'] = $this->errorDetail;
+        }
+
+        if (null !== $this->errorType) {
+            $res['ErrorType'] = $this->errorType;
         }
 
         if (null !== $this->importTagKey) {
@@ -171,6 +191,14 @@ class applicationGroups extends Model
 
         if (isset($map['Description'])) {
             $model->description = $map['Description'];
+        }
+
+        if (isset($map['ErrorDetail'])) {
+            $model->errorDetail = $map['ErrorDetail'];
+        }
+
+        if (isset($map['ErrorType'])) {
+            $model->errorType = $map['ErrorType'];
         }
 
         if (isset($map['ImportTagKey'])) {